requirements.txt
slack_bolt
slack_sdk
psycopg2-binary
python-dotenv
